backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg[binary]
python-jose[cryptography]
pydantic-settings
python-dotenv
email-validator
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-multipart
cloudinary